requirements.txt
accelerate==1.12.0
affine==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bitsandbytes==0.48.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.2.0
click-plugins==1.1.1.2
cligj==0.7.2
clip==0.2.0
cloudpickle==3.1.2
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.4.1
diff-match-patch==20241021
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
emoji==2.15.0
evaluate==0.4.6
faiss-gpu-cu12==1.13.0
Faker==37.12.0
fastapi==0.122.0
fastuuid==0.14.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
genson==1.3.0
geopandas==1.1.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.36.0
grpcio==1.76.0
guardrails-ai==0.7.0
guardrails-api-client==0.4.0
guardrails_hub_types==0.0.4
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lark==1.3.1
lazy_imports==1.1.0
litellm==1.80.0
llama_cpp_python==0.3.16
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mcp==1.3.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.9.79
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.2
openai==1.109.1
openai-clip==1.0.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outlines==1.2.9
outlines_core==0.2.11
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
ply==3.11
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydash==8.0.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyogrio==0.12.1
pyparsing==3.2.5
pypdf==6.4.0
PyPika==0.48.9
pyproj==3.7.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
rasterio==1.4.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
rstr==3.2.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semver==3.0.4
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1+cu129
torchaudio==2.9.1+cu129
torchvision==0.24.1+cu129
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.21
unstructured-client==0.42.4
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
Wikipedia-API==0.8.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
